+++ lib/src/tools/launch_exception.dart
@@ -1,0 +1,8 @@
+class LaunchException implements Exception {
+  final String message;
+
+  LaunchException(this.message);
+
+  @override
+  String toString() => 'LaunchException: $message';
+}